Got it working!
The system works as follows (see __init__.py in /v4l as an example):

grabImage.so defines two functions, grab_image and free_image; import it as you would any other module.  grab_image takes three arguments, the name of the camera device on the system (e.g., "/dev/video0"), the width of the grab desired, and the height.  It returns a list of bytes.  I'm pretty sure that every three list entries is an RGB tuple; it's actually a direct mapping of the memory that the C program grabs from the camera.  When you're done with the image file, call free_image(x) (x being the list that grab_image returned) to free the memory that was allocated to hold the image data.

If you make any changes to the .c or .h files, call make to rebuild grabImage.so.  The various Makefiles are required to build the C files with python.