This file implements Fahlman's cascade correlation algorithm as described in his paper "The Cascade-Correlation Learning Architecture". A new class called CascadeCorNet inherits from Network.  There are a few functions defined at the top of the file that will eventually be moved into the class.  Currently, CascadeCorNet does not work.  It adds as many hidden layers as it is allowed to.  It seems to be training successfully, however.  It just doesn't stop when the error drops.
Here is a list of some of the important functions defined for CascadeCorNets.  Docstrings for these functions are not yet up to date, nor are internal
comments.

__init__ :  Turns on quickprop by default.

train : Responsible for the actual training.  This is the function that users of CascadeCorNet should
call.  It breaks down training into the output training phase, the candidate training phase, and the
recruiting phase.  train consumes a maximum number of hidden units to add.

trainCandidates : This function contains most of the new code for cascade correlation.  It trains the
candidate layer to correlate with the output error.  Most of the variable names are inspired by the equations
in Fahlman's paper.

computeDataFromProp : Used in trainCandidates to calculate various quantities that are based on propagation of all the input
through the network and that need not be recomputed when candidate weights change.

computeChangingDataFromProp :  Also used in trainCandidates, but is called before every change in candidate weights in order
to calculate things like the candidate activation values for each pattern and the net input to the candidate layer.

computeS_c : This function calculates the correlation quantity Fahlman calls 'S' in his paper.  This is the score we want
to maximize for the candidate layer units.

compute_dSdw : Calculates the partial derivative of of S with respect to a weight.  The docstrings for these functions
need to be written so important information about the shape of the arrays they consume and return is recorded somewhere.

trainOutputs : This function is mostly copied from Network::train.  Some small changes are required to detect stagnation.

recruit: Copied and unchanged from IncrementalNetwork.

This file needs a lot of work before it will work nicely.  It also needs a lot of comment rewriting and cleaning up.